no change

+++ Test.java
@@ -15,6 +15,6 @@ public class Test
    
    public static float divider(int no1, int no2)
    {
-      return no1/no2;
+      return no1/(float) no2;
    }
 }
